+++ src/Block/UrlListProvider/AbstractUrlListProvider.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Sobak\Scrawler\Block\UrlListProvider;
 
-use GuzzleHttp\Psr7\Response;
+use Psr\Http\Message\ResponseInterface;
 use Sobak\Scrawler\Client\Response\Elements\Url;
 
 abstract class AbstractUrlListProvider implements UrlListProviderInterface
@@ -15,7 +15,7 @@ abstract class AbstractUrlListProvider implements UrlListProviderInterface
     /** @var Url */
     protected $currentUrl;
 
-    /** @var Response */
+    /** @var ResponseInterface */
     protected $response;
 
     public function setBaseUrl(Url $url): void
@@ -28,7 +28,7 @@ abstract class AbstractUrlListProvider implements UrlListProviderInterface
         $this->currentUrl = $url;
     }
 
-    public function setResponse(Response $response): void
+    public function setResponse(ResponseInterface $response): void
     {
         $this->response = $response;
     }