+++ src/Block/UrlListProvider/EmptyUrlListProvider.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Sobak\Scrawler\Block\UrlListProvider;
 
-use GuzzleHttp\Psr7\Response;
+use Psr\Http\Message\ResponseInterface;
 use Sobak\Scrawler\Client\Response\Elements\Url;
 
 class EmptyUrlListProvider implements UrlListProviderInterface
@@ -24,7 +24,7 @@ class EmptyUrlListProvider implements UrlListProviderInterface
         // Do nothing...
     }
 
-    public function setResponse(Response $response): void
+    public function setResponse(ResponseInterface $response): void
     {
         // Do nothing...
     }